Split MutationVariant.cs, Moved TestHostRunnerFactory.cs
