+++ Faultify.TestRunner/TestRun/MutationVariant.cs
@@ -1,35 +1,9 @@
-﻿using System.Collections.Generic;
-using System.Reflection.Metadata;
+﻿using System.Reflection.Metadata;
 using Faultify.Analyze.AssemblyMutator;
 using Faultify.Analyze.Mutation;
 
 namespace Faultify.TestRunner.TestRun
 {
-    /// <summary>
-    ///     Single mutation variant that serves as an identifier for a particular possible mutation.
-    /// </summary>
-    public readonly struct MutationVariantIdentifier
-    {
-        public MutationVariantIdentifier(
-            HashSet<string> testNames,
-            string memberName,
-            int mutationId,
-            int mutationGroupId
-        )
-        {
-            TestCoverage = testNames;
-            MemberName = memberName;
-            MutationId = mutationId;
-            MutationGroupId = mutationGroupId;
-        }
-
-        public int MutationId { get; }
-        public int MutationGroupId { get; }
-        public string MemberName { get; }
-        public HashSet<string> TestCoverage { get; }
-    }
-
-
     /// <summary>
     ///     Single mutation variant which contains the executable mutation.
     /// </summary>
